requirements.txt
aspose_words==22.10.0
datefinder==0.7.3
lxml==4.9.1
numpy==1.20.0
pandas==1.2.1
pdfminer.six==20221105
Pillow==8.1.0
pyodbc==4.0.34
pypyodbc==1.3.6
PyPDF2==2.11.1
python_dateutil==2.8.1
tqdm==4.56.0
altgraph==0.17.3
aspose-words==22.10.0
cffi==1.15.1
chardet==5.0.0
charset-normalizer==3.0.1
cryptography==38.0.3
future==0.18.2
pefile==2022.5.30
ply==3.11
pycparser==2.21
pycryptodome==3.15.0
pyinstaller==5.6.2
pyinstaller-hooks-contrib==2022.13
python-dateutil==2.8.1
pytz==2022.6
pywin32-ctypes==0.2.0
regex==2022.10.31
six==1.16.0
sortedcontainers==2.4.0
typing_extensions==4.4.0
